prettier fixes

+++ src/test/integration/pools/FusePool156test.t.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity >=0.8.0;
 
-import { BaseDSTest } from "../../utils/BaseDSTest.sol";
+import {BaseDSTest} from "../../utils/BaseDSTest.sol";
 import "forge-std/Vm.sol";
 
 import {ERC20} from "solmate/tokens/ERC20.sol";
@@ -14,23 +14,27 @@ import "../../interfaces/IPlugin.sol";
 
 contract FusePool156Test is BaseDSTest {
     FuseAdmin fuseAdmin = FuseAdmin(0xa731585ab05fC9f83555cf9Bff8F58ee94e18F85);
-    
+
     // pool 156 comptroller
-    Comptroller comptroller = Comptroller(0x07cd53380FE9B2a5E64099591b498c73F0EfaA66);
+    Comptroller comptroller =
+        Comptroller(0x07cd53380FE9B2a5E64099591b498c73F0EfaA66);
     address multisig = 0x5eA4A9a7592683bF0Bc187d6Da706c6c4770976F;
     address user = 0xB290f2F3FAd4E540D0550985951Cdad2711ac34A;
 
     address cvxcrvCRV = 0x9D0464996170c6B9e75eED71c68B99dDEDf279e8;
     address cvxFXSFXS = 0xF3A43307DcAFa93275993862Aae628fCB50dC768;
     address rethstethCRV = 0x447Ddd4960d9fdBF6af9a790560d0AF76795CB08;
-    
+
     ERC20 cvx = ERC20(0x4e3FBD56CD56c3e72c1403e103b45Db9da5B9D2B);
     ERC20 crv = ERC20(0xD533a949740bb3306d119CC777fa900bA034cd52);
     ERC20 fxs = ERC20(0x3432B6A60D23Ca0dFCa7761B7ab56459D9C964D0);
 
-    FlywheelCore cvxFlywheelCore = FlywheelCore(0x18B9aE8499e560bF94Ef581420c38EC4CfF8559C);
-    FlywheelCore crvFlywheelCore = FlywheelCore(0x65DFbde18D7f12a680480aBf6e17F345d8637829);
-    FlywheelCore fxsFlywheelCore = FlywheelCore(0x30E9A1Bc6A6a478fC32F9ac900C6530Ad3A1616F);
+    FlywheelCore cvxFlywheelCore =
+        FlywheelCore(0x18B9aE8499e560bF94Ef581420c38EC4CfF8559C);
+    FlywheelCore crvFlywheelCore =
+        FlywheelCore(0x65DFbde18D7f12a680480aBf6e17F345d8637829);
+    FlywheelCore fxsFlywheelCore =
+        FlywheelCore(0x30E9A1Bc6A6a478fC32F9ac900C6530Ad3A1616F);
 
     function setUp() public {
         hevm.label(address(fuseAdmin), "fuseAdmin");
@@ -50,30 +54,41 @@ contract FusePool156Test is BaseDSTest {
         require(cvxCRVf.mint(100e18) == 0, "mint failed");
 
         deal(rethstethCRV, user, 100e18);
-        CErc20 rethstethCRVf = CErc20(0xD88B2E6304d1827e22D2ACC2FbCeD836cd439b85);
+        CErc20 rethstethCRVf = CErc20(
+            0xD88B2E6304d1827e22D2ACC2FbCeD836cd439b85
+        );
         require(rethstethCRVf.mint(100e18) == 0, "mint failed");
 
         // cvxFXSFXSf
-        uint accrue = cvxFlywheelCore.accrue(ERC20(address(cvxFXSf)), user);
+        uint256 accrue = cvxFlywheelCore.accrue(ERC20(address(cvxFXSf)), user);
         require(accrue > 0, "cvxFlywheel accrue");
-        uint accrued = cvxFlywheelCore.rewardsAccrued(user);
-        uint prebalance = cvx.balanceOf(user);
+        uint256 accrued = cvxFlywheelCore.rewardsAccrued(user);
+        uint256 prebalance = cvx.balanceOf(user);
         cvxFlywheelCore.claimRewards(user);
-        require(cvx.balanceOf(user) == prebalance + accrued, "cvxFlywheel claimRewards");
+        require(
+            cvx.balanceOf(user) == prebalance + accrued,
+            "cvxFlywheel claimRewards"
+        );
 
         accrue = crvFlywheelCore.accrue(ERC20(address(cvxFXSf)), user);
         require(accrue > 0, "crvFlywheel accrue");
         accrued = crvFlywheelCore.rewardsAccrued(user);
         prebalance = crv.balanceOf(user);
         crvFlywheelCore.claimRewards(user);
-        require(crv.balanceOf(user) == prebalance + accrued, "crvFlywheel claimRewards");
+        require(
+            crv.balanceOf(user) == prebalance + accrued,
+            "crvFlywheel claimRewards"
+        );
 
         accrue = fxsFlywheelCore.accrue(ERC20(address(cvxFXSf)), user);
         require(accrue > 0, "fxsFlywheel accrue");
         accrued = fxsFlywheelCore.rewardsAccrued(user);
         prebalance = fxs.balanceOf(user);
         fxsFlywheelCore.claimRewards(user);
-        require(fxs.balanceOf(user) == prebalance + accrued, "fxsFlywheel claimRewards");
+        require(
+            fxs.balanceOf(user) == prebalance + accrued,
+            "fxsFlywheel claimRewards"
+        );
 
         // cvxCRVf
         hevm.warp(block.timestamp + 10);
@@ -82,14 +97,20 @@ contract FusePool156Test is BaseDSTest {
         accrued = cvxFlywheelCore.rewardsAccrued(user);
         prebalance = cvx.balanceOf(user);
         cvxFlywheelCore.claimRewards(user);
-        require(cvx.balanceOf(user) == prebalance + accrued, "cvxFlywheel claimRewards");
+        require(
+            cvx.balanceOf(user) == prebalance + accrued,
+            "cvxFlywheel claimRewards"
+        );
 
         accrue = crvFlywheelCore.accrue(ERC20(address(cvxCRVf)), user);
         require(accrue > 0, "crvFlywheel accrue");
         accrued = crvFlywheelCore.rewardsAccrued(user);
         prebalance = crv.balanceOf(user);
         crvFlywheelCore.claimRewards(user);
-        require(crv.balanceOf(user) == prebalance + accrued, "crvFlywheel claimRewards");
+        require(
+            crv.balanceOf(user) == prebalance + accrued,
+            "crvFlywheel claimRewards"
+        );
 
         // rethstethCRVf
         hevm.warp(block.timestamp + 10);
@@ -98,17 +119,26 @@ contract FusePool156Test is BaseDSTest {
         accrued = cvxFlywheelCore.rewardsAccrued(user);
         prebalance = cvx.balanceOf(user);
         cvxFlywheelCore.claimRewards(user);
-        require(cvx.balanceOf(user) == prebalance + accrued, "cvxFlywheel claimRewards");
+        require(
+            cvx.balanceOf(user) == prebalance + accrued,
+            "cvxFlywheel claimRewards"
+        );
 
         accrue = crvFlywheelCore.accrue(ERC20(address(rethstethCRVf)), user);
         require(accrue > 0, "crvFlywheel accrue");
         accrued = crvFlywheelCore.rewardsAccrued(user);
         prebalance = crv.balanceOf(user);
         crvFlywheelCore.claimRewards(user);
-        require(crv.balanceOf(user) == prebalance + accrued, "crvFlywheel claimRewards");
+        require(
+            crv.balanceOf(user) == prebalance + accrued,
+            "crvFlywheel claimRewards"
+        );
 
         require(cvxFXSf.redeem(cvxFXSf.balanceOf(user)) == 0, "cvxFXS redeem");
         require(cvxCRVf.redeem(cvxCRVf.balanceOf(user)) == 0, "cvxCRV redeem");
-        require(rethstethCRVf.redeem(rethstethCRVf.balanceOf(user)) == 0, "rethstethCRV redeem");
+        require(
+            rethstethCRVf.redeem(rethstethCRVf.balanceOf(user)) == 0,
+            "rethstethCRV redeem"
+        );
     }
 }
